requirements.txt
pytest==8.4.2
allure-pytest==2.15.3
requests==2.31.0
openpyxl==3.1.2
pytest-html==4.2.0